Linux compilation warning cleanup. Storage: change readStorage/writeStorate to take unsinged int for size (may break compilation if custom Storage interfaces are used).

+++ extras/porting/linux/supla/parser/simple.cpp
@@ -28,7 +28,7 @@ Supla::Parser::Simple::~Simple() {
 
 double Supla::Parser::Simple::getValue(const std::string &key) {
   int index = keys[key];
-  if (index < 0 || index >= values.size()) {
+  if (index < 0 || static_cast<size_t>(index) >= values.size()) {
     valid = false;
     return 0;
   }
@@ -46,7 +46,7 @@ double Supla::Parser::Simple::getValue(const std::string &key) {
 std::variant<int, bool, std::string> Supla::Parser::Simple::getStateValue(
     const std::string &key) {
   int index = keys[key];
-  if (index < 0 || index >= values.size()) {
+  if (index < 0 || static_cast<size_t>(index) >= values.size()) {
     valid = false;
     return 0;
   }
